Move worker actions & fetching to router package
